+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 * Remove special chars from tokens
 * Aquos contact feedback with name and booking date
+* Due to problem with Aquos webservice, the contacts are now sent with a cron every 5 minutes, and not through the gravity form submission
 
 ### 1.5.0: June 15th, 2021
 * Fix filter bookings query